New web using Quarto
While I am thankful to the creator of the original Hugo theme that I was using beforehand, I felt it would be easier to maintain the project using the Quarto framework. The overall project is much more lightweight, though the website files are slightly larger. Personally, I am a big fan of Quarto and I also use it for data science, which means that it's much more familiar to me as I am not a web developer.